4.21*: bloglist expansion, step 9

+++ part4/blogList/controllers/blogs.js
@@ -51,7 +51,20 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const token = request.token;
+
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid"});
+  }
+  const tokenUserId = decodedToken.id;
+  const blog = await Blog.findById(request.params.id);
+  const blogUserId = blog.user.toString();
+  if(!(tokenUserId === blogUserId)) {
+    return response.status(401).json({ error: "wrong user"});
+  } 
   await Blog.findByIdAndRemove(request.params.id);
+  
   response
     .status(204)
     .end();
